按钮收到 touch_move 消息时，转发给父窗口。 触屏消息坐标和滑动速度是负数的处理。

+++ component/gui/gdd/gdd_input.c
@@ -197,8 +197,8 @@ void GDD_HmiInput(void)
                         Touch_Msg = MSG_TOUCH_UP;
                 }
                 PostMessage(hwnd, Touch_Msg,
-                            (TouchMsg->MoveY << 16) | TouchMsg->MoveX,
-                            (pt.y << 16) | pt.x);
+                            ((u16)(TouchMsg->MoveY) << 16) | (u16)(TouchMsg->MoveX),
+                            ((u16)(pt.y) << 16) | (u16)(pt.x));
 
 
 //
