+++ component/include/hmi-input.h
@@ -109,7 +109,7 @@ struct SingleTouchMsg
     //指明触摸事件对应的屏幕，NULL表示默认显示器，如果只有一个屏也可设为NULL。
     struct DisplayObj *display;
     s32 x,y,z;              //x,y表示触摸位置，z>0标志触摸压力，0标志未触摸
-    s32 MoveX,MoveY;        //每ms移动的微米数
+    s16 MoveX,MoveY;        //每ms移动的微米数
 };
 
 //多点触摸屏
